pyproject.toml
[project]
name = "al1s-bot"
version = "1.1.0"
description = "一个功能完整的Telegram聊天机器人，支持OpenAI聊天、RAG知识增强、角色定制和图片搜索"
readme = "README.md"
license = "MIT"
authors = [
    {name = "TexasOct", email = "aecom233@gmail.com"}
]
keywords = ["telegram", "bot", "ai", "rag", "chatbot", "openai"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.13"
dependencies = [
  "python-telegram-bot>=22.3",
  "aiohttp>=3.8.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "loguru>=0.7.0",
  "beautifulsoup4>=4.12.0",
  "pillow>=11.3.0",
  "requests>=2.31.0",
  "PicImageSearch>=3.12.9",
  "pyyaml>=6.0.2",
  "openai>=1.105.0",
  "mcp[cli]>=1.0.0",
  "faiss-cpu>=1.7.4",
  "numpy>=2.0.0",
  "sentence-transformers>=5.1.0",
  "langchain>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-openai>=0.2.0",
  "langgraph>=0.2.0",
  "langchain-text-splitters>=0.3.0",
  "tiktoken>=0.7.0",
  "langchain-huggingface>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/texasoct/al1s"
Repository = "https://github.com/texasoct/al1s"
Issues = "https://github.com/texasoct/al1s/issues"
Documentation = "https://github.com/texasoct/al1s#README.md"

[tool.uv]
dev-dependencies = [
  "pytest>=8.4.1",
  "pytest-asyncio>=1.1.0",
  "black>=25.1.0",
  "isort>=6.0.1",
  "flake8>=7.3.0",
]

[[tool.uv.index]]
url = "https://mirrors.ustc.edu.cn/pypi/simple/"
default = true